CRBN-338 When building DAG, we shouldn't blindly lock. Use int instead of double for default timeout

+++ src/Miningcore/Crypto/Hashing/Ethash/Dag.cs
@@ -22,7 +22,7 @@ namespace Miningcore.Crypto.Hashing.Ethash
 
         private IntPtr handle = IntPtr.Zero;
         private static readonly Semaphore sem = new Semaphore(1, 1);
-        private const double DefaultLockTimeoutInMinutes = 40;
+        private const int DefaultLockTimeoutInMinutes = 40;
 
         public DateTime LastUsed { get; set; }
